WIP pundit authorization for posts

+++ app/controllers/topics/posts_controller.rb
@@ -1,7 +1,7 @@
 class Topics::PostsController < ApplicationController
   def show
     @topic = Topic.find(params[:topic_id])
-    authorize! :read, @topic, message: "You need to be user to do that."
+    authorize @topic
     @post = Post.find(params[:id])
     @comments = @post.comments
     @comment = Comment.new
@@ -10,20 +10,20 @@ class Topics::PostsController < ApplicationController
   def new
     @topic = Topic.find(params[:topic_id])
     @post = Post.new
-    authorize! :create, Post, message: "You need to be a member to create a new post."
+    authorize @post
   end
 
   def edit
     @topic = Topic.find(params[:topic_id])
     @post = Post.find(params[:id])
-    authorize! :edit, @post, message: "You need to own the post to edit it."
+    authorize @post
   end
   
   def create
     @topic = Topic.find(params[:topic_id])
     @post = current_user.posts.build(post_params)
     @post.topic = @topic
-    authorize! :create, @post, message: "You need to be signed up to do that."
+    authorize @post
     if @post.save
       flash[:notice] = "Post was saved."
       redirect_to [@topic, @post]
@@ -36,7 +36,7 @@ class Topics::PostsController < ApplicationController
   def update
     @topic = Topic.find(params[:topic_id])
     @post = Post.find(params[:id])
-    authorize! :update, @post, message: "You need to own the post to edit it."
+    authorize @post
     if @post.update_attributes(post_params)
       flash[:notice] = "Post was updated."
       redirect_to [@topic, @post]
@@ -53,7 +53,7 @@ class Topics::PostsController < ApplicationController
     @comment = Comment.new
 
     title = @post.title
-    authorize! :destroy, @post, message: "You need to own the post to delete it."
+    authorize @post
     if @post.destroy
       flash[:notice] = "\"#{title}\" was deleted successfully."
       redirect_to @topic
